renamed model in line with new naming cocnvention
